% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_gibbs.R
\name{count_rej}
\alias{count_rej}
\title{Counting the rejections}
\usage{
count_rej(res_list)
}
\arguments{
\item{res_list}{list of the results of the gibbs sampling procedure}
}
\value{
number of total rejections
}
\description{
This function summarizes the value for the rejected samples
}
