% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{countRejs}
\alias{countRejs}
\title{Counting the rejections}
\usage{
countRejs(resList)
}
\arguments{
\item{resList}{list of the results of the gibbs sampling procedure}
}
\value{
number of total rejections
}
\description{
This function summarizes the value for the rejected samples
}
