% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtmvnorm_fast.R
\name{dtmvnorm_fast}
\alias{dtmvnorm_fast}
\title{Fast density of truncated multivariate Gaussian}
\usage{
dtmvnorm_fast(x, mean = rep(0, nrow(sigma)),
  sigma = diag(length(mean)), lower = rep(-Inf, length = length(mean)),
  upper = rep(Inf, length = length(mean)), log = FALSE)
}
\arguments{
\item{x}{matrix of observations where the rows are individual observations and the columns are the dimension}

\item{mean}{mean of the density to be evaluated}

\item{sigma}{covariance matrix of the density to be evaluated}

\item{lower}{lower bounds of the truncation}

\item{upper}{upper bounds of the truncation}

\item{log}{whether to return log probabilities}
}
\value{
value of density
}
\description{
This function computes the density of a truncated multivariate Gaussian faster than mvtnorm::dtmvnorm and use mvnfast to evaluate the density
}
