% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_sampling.R, R/utils.R
\name{dtmvnorm_impsamp}
\alias{dtmvnorm_impsamp}
\title{Truncated univariate Gaussian density}
\usage{
dtmvnorm_impsamp(X, mn, sig, constr, num_samp = 1e+06)

dtmvnorm_impsamp(X, mn, sig, constr, num_samp = 1e+06)
}
\arguments{
\item{X}{a vector of observations where each column is the independent observations}

\item{mn}{mean of the Gaussian density}

\item{sig}{covariance matrix of the Gaussian density}

\item{constr}{a function that describes the domain or constraint of the samples}

\item{num_samp}{number of samples in the important sampling procedure}

\item{X}{a vector of observations where each column is the independent observations}

\item{mn}{mean of the Gaussian density}

\item{sig}{covariance matrix of the Gaussian density}

\item{constr}{a function that describes the domain or constraint of the samples}

\item{num_samp}{number of samples in the important sampling procedure}
}
\value{
value of the density

value of the density
}
\description{
This function evaluates the truncated univariate Gaussian density using an important sampling procedure as the normalizing constant

This function evaluates the truncated univariate Gaussian density using an important sampling procedure as the normalizing constant
}
\examples{
X <- rnorm(5, mean=0, sd=1)
constr <- function(X, u1 = lower, u2 = upper) { X >= u1 & X <= u2 }
dtmvnorm_impsamp(X, mn, sig, constr, num_samp=1e6)
X <- rnorm(5, mean=0, sd=1)
constr <- function(X, u1 = lower, u2 = upper) { X >= u1 & X <= u2 }
dtmvnorm_impsamp(X, mn, sig, constr, num_samp=1e6)
}
