% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dtmvnorm_impsamp_mv}
\alias{dtmvnorm_impsamp_mv}
\title{Truncated Multivariate Gaussian density}
\usage{
dtmvnorm_impsamp_mv(X, mn, sig, constr, num_samp = 10000, log = T)
}
\arguments{
\item{X}{a vector of observations where each row is the dimension and each column is the independent observations}

\item{mn}{mean of the multivariate Gaussian density}

\item{sig}{covariance matrix of the multivariate Gaussian density}

\item{constr}{a function that describes the domain or constraint of the samples}

\item{num_samp}{number of samples in the important sampling procedure}

\item{log}{whether log value is returned (default TRUE)}
}
\value{
value of density
}
\description{
This function evaluates the truncated multivariate Gaussian density using an important sampling procedure as the normalizing constant
}
\examples{
genconst <- function(u1, u2, l1, l2) { in_set <- function(X) {
 if (nrow(X) == 1) X = t(X)
 X[1,] > l1 & X[1,] < u1 & X[2,] > l2 & X[2,] < u2 & X[3,] > l1 &
 X[3,] < u1 & X[4,] > l2 & X[4,] < u2 }}
constr  <- genconst(u1 = 0, u2 = 0, l1 = 1, l2 = 1)
dtmvnorm_impsamp_mv(X = mvnfast::rmvn(5, mu = rep(0,4), sigma = diag(0.1, 4)), mn = rep(0, 4), sig = diag(0.1,4), constr, num_samp = 1000, log = F)
}
