% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_pred_func.R
\name{edit_log_prob}
\alias{edit_log_prob}
\title{Edit log probability}
\usage{
edit_log_prob(X)
}
\arguments{
\item{X}{vector of probabilities.}
}
\value{
vector of substituted log probabilities
}
\description{
This function edits the log probability for underflow and overflow. Substitute Inf to -Inf and NaN to -Inf.
}
