% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{editLogProb}
\alias{editLogProb}
\title{Edit log probability}
\usage{
editLogProb(X)
}
\arguments{
\item{X}{vector of probabilities.}
}
\value{
vector of substituted log probabilities
}
\description{
This function edits the log probability for underflow and overflow. Substitute Inf to -Inf and NaN to -Inf.
}
