% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{editPredProb}
\alias{editPredProb}
\title{Edit underflow of predictive probability}
\usage{
editPredProb(X)
}
\arguments{
\item{X}{vector of probabilities}
}
\description{
This function edits the underflow of the predictive probability function for truncated Gaussian
}
