% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{editProb}
\alias{editProb}
\title{Edit probability}
\usage{
editProb(X)
}
\arguments{
\item{X}{vector of probabilities}
}
\value{
vector of fixed probabilities
}
\description{
This function fixes the underflow and overflow of probability output with 0's
}
