% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_fig_for_paper.R
\name{fig_error_post}
\alias{fig_error_post}
\title{Algorithm for Error in the posterior}
\usage{
fig_error_post(test_pred)
}
\arguments{
\item{test_pred}{predictive probability for the test set}
}
\value{
plot of error for the predictive density on the test set
}
\description{
This function generates an error plot of the posterior distribution
}
