% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_fig_for_paper.R
\name{fig_numrejects}
\alias{fig_numrejects}
\title{Number of rejection plot}
\usage{
fig_numrejects(count_bins, mix, thr_l)
}
\arguments{
\item{count_bins}{number of rejections in each bins}

\item{mix}{one of "tmogt" and "motg"}

\item{thr_l}{lenght of the threshold vector}
}
\value{
histogram for the number of rejections in each threshold
}
\description{
This function generates a plot of the number of rejections
}
