% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_fig_for_paper.R
\name{fig_pred_prob_gauss}
\alias{fig_pred_prob_gauss}
\title{Plot predictive density for Gaussian}
\usage{
fig_pred_prob_gauss(test_pred, thr, mu_r = 0, labs = T)
}
\arguments{
\item{test_pred}{predictive probability at the grid}

\item{thr}{maximum number of rejected proposals for each observation}

\item{mu_r}{mode of Gaussian}

\item{labs}{whether labels should be included in the figure}
}
\value{
posterior predictive density plot
}
\description{
This function generates the posterior predictive density for the Gaussian case
}
