% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_fig_for_paper.R
\name{func_perf_plot_gauss}
\alias{func_perf_plot_gauss}
\title{Algorithm performance in Gaussian distribution}
\usage{
func_perf_plot_gauss(time_all_result, ppred_all_result, thr, labs)
}
\arguments{
\item{time_all_result}{list of times for all result}

\item{ppred_all_result}{list of prediction for all result}

\item{thr}{vector of maximum number of rejections}

\item{labs}{whether a legend should be included in the figures}
}
\value{
plot of the performance on a Beta distribution
}
\description{
This function generates a plot for the performance on a Gaussian distribution
}
