% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gen_mog}
\alias{gen_mog}
\title{Mixture of Gaussians}
\usage{
gen_mog(K = 2, param_list, N = 100, dm)
}
\arguments{
\item{K}{number of clusters (default 2)}

\item{param_list}{list containing weight (wt), mean (mu), and covariance (cv)}

\item{N}{number of samples}

\item{dm}{dimensions of Gaussian}
}
\value{
list with Gaussian samples and component assignment
}
\description{
This function generates mixtures of gaussians samples
}
