% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{genMog}
\alias{genMog}
\title{Mixture of Gaussians}
\usage{
genMog(K = 2, paramList, N = 100, dm)
}
\arguments{
\item{K}{number of clusters (default 2)}

\item{paramList}{list containing weight (wt), mean (mu), and covariance (cv)}

\item{N}{number of samples}

\item{dm}{dimensions of Gaussian}
}
\value{
list with Gaussian samples and component assignment
}
\description{
This function generates mixtures of gaussians samples
}
