% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gibbSamplingOri}
\alias{gibbSamplingOri}
\title{Gibbs sampling algorithm}
\usage{
gibbSamplingOri(K, ipdata, thr = Inf, burnIn, numIter, constr, mix,
  lower, upper, dataType)
}
\arguments{
\item{K}{maximum number of clusters}

\item{ipdata}{data matrix where columns are individual observations and rows are dimensions}

\item{thr}{maximum number of rejections per observations}

\item{burnIn}{number of burn-in for the chain. Must be less than number of iterations}

\item{numIter}{total number of iterations}

\item{constr}{function for the constraints}

\item{mix}{one of "tmog" and "motg"}

\item{lower}{lower constraint for unit square}

\item{upper}{upper constriant for unit square}

\item{dataType}{one of "gauss", "beta", "flowcyto", "crime", "bvgauss"}
}
\value{
a list of results consisting of weights, mean, and covariance matrix for every iterations, along with the rejected proposals
}
\description{
This computes the Gibbs sampling procedure
}
