% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_func.R
\name{give_ind}
\alias{give_ind}
\title{Index of rejections}
\usage{
give_ind(vl, c_ind)
}
\arguments{
\item{vl}{a vector of TRUE/FALSE indicating whether the samples are inside or outside the constraints}

\item{c_ind}{number of last accepted observations}
}
\description{
This function gives the rejected samples indices associated to the observation it belongs to
}
