% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{giveInd}
\alias{giveInd}
\title{Index of rejections}
\usage{
giveInd(vl, cind)
}
\arguments{
\item{vl}{a vector of TRUE/FALSE indicating whether the samples are inside or outside the constraints}

\item{cInd}{number of last accepted observations}
}
\description{
This function gives the rejected samples indices associated to the observation it belongs to
}
