% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{import_sampling_mv}
\alias{import_sampling_mv}
\title{Important Sampling for Multivariate variables}
\usage{
import_sampling_mv(constr, mn, cv, mu_prop, sig_prop, num_obs)
}
\arguments{
\item{constr}{a function that describes the domain or constraint of the observations}

\item{mn}{mean of the observations to be generated}

\item{cv}{covariance matrix of the observations to be generated}

\item{mu_prop}{mean of the proposal distribution}

\item{sig_prop}{covariance matrix of the proposal distribution}

\item{num_obs}{number of observations to be generated}
}
\value{
value of integrals
}
\description{
This function evaluates integrals of a multivariate random variables using important sampling method with a multivariate Gaussian proposal
}
\examples{
genconst <- function(u1, u2, l1, l2) { in_set <- function(X) {
 if (nrow(X) == 1) X = t(X)
 X[1,] > l1 & X[1,] < u1 & X[2,] > l2 & X[2,] < u2 & X[3,] > l1 &
 X[3,] < u1 & X[4,] > l2 & X[4,] < u2 }}
constr  <- genconst(u = 0, u = 0, l1 = 1, l2 = 1)
import_sampling_mv(constr, rep(1/2, 4), diag(0.1,4), rep(0,4), diag(0.3,4), 100)
}
