% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_sampling.R
\name{import_sampling_norm}
\alias{import_sampling_norm}
\title{Normalized Important Sampling function}
\usage{
import_sampling_norm(constr, mn, cv, num_obs)
}
\arguments{
\item{constr}{a function that describes the domain or constraint of the observations}

\item{mn}{mean of the observations to be generated}

\item{cv}{covariance matrix of the observations to be generated}

\item{num_obs}{number of observations to be generated}
}
\value{
value of integrals
}
\description{
This function evaluates integrals using a normalized important sampling method with a Gaussian distribution proposal
}
\examples{
constr <- function(X, u1 = lower, u2 = upper) { X >= u1 & X <= u2 }
important_sampling_norm(constr, 1/2, 0.1, 100)

}
