% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{importanceSampling}
\alias{importanceSampling}
\title{Important Sampling function}
\usage{
importanceSampling(constr, mn, cv, numObs)
}
\arguments{
\item{constr}{a function that describes the domain or constraint of the observations}

\item{mn}{mean of the observations to be generated}

\item{cv}{covariance matrix of the observations to be generated}

\item{numObs}{number of observations to be generated}
}
\value{
value of integrals
}
\description{
This function evaluates integrals using an important sampling method with a Gaussian distribution proposal
}
\examples{
constr <- function(X, u1 = lower, u2 = upper) { X >= u1 & X <= u2 }
important_sampling(constr, 1/2, 0.1, 100)

}
