% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_func.R
\name{impute_gauss}
\alias{impute_gauss}
\title{Imputation of rejected samples}
\usage{
impute_gauss(mu, cv, N, constr, thr, dm, ceil_rej)
}
\arguments{
\item{mu}{mean parameters}

\item{cv}{covariance matrix}

\item{N}{total number of observations}

\item{constr}{function for the constraint of the space}

\item{thr}{maximum number of rejections per observation}

\item{dm}{dimension of observations}

\item{ceil_rej}{max number of observations in total}
}
\description{
This function samples the rejected proposals for the Mixture of Truncated Gaussians (MoTG) model
}
