% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_func.R
\name{impute_rej}
\alias{impute_rej}
\title{Imputation of rejected samples}
\usage{
impute_rej(K, params, N, constr, thr, dm, ceil_rej)
}
\arguments{
\item{K}{total number of clusters}

\item{params}{list of parameters which includes the weight (wt), mean (mu), and covariance matrix (cv)}

\item{N}{total number of observations}

\item{constr}{function for the constraint of the space}

\item{thr}{maximum number of rejections per observation}

\item{dm}{dimension of observations}

\item{ceil_rej}{max number of observations in total}
}
\description{
This function samples the rejected proposals for the Truncated Mixture of Gaussian (TMoG) model
}
