% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{imputeGauss}
\alias{imputeGauss}
\title{Imputation of rejected samples}
\usage{
imputeGauss(mu, cv, N, constr, thr, ceilRej)
}
\arguments{
\item{mu}{mean parameters}

\item{cv}{covariance matrix}

\item{N}{total number of observations}

\item{constr}{function for the constraint of the space}

\item{thr}{maximum number of rejections per observation}

\item{ceilRej}{max number of observations in total}
}
\value{
rejected samples and indices to which observation it belongs to
}
\description{
This function samples the rejected proposals for the Mixture of Truncated Gaussians (MoTG) model
}
