% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{imputeRejections}
\alias{imputeRejections}
\title{Imputation of rejected samples}
\usage{
imputeRejections(K, params, N, constr, thr, ceilRej)
}
\arguments{
\item{K}{total number of clusters}

\item{params}{list of parameters which includes the weight (wt), mean (mu), and covariance matrix (cv)}

\item{N}{total number of observations}

\item{constr}{function for the constraint of the space}

\item{thr}{maximum number of rejections per observation}

\item{ceilRej}{max number of observations in total}
}
\value{
rejected samples and indices to which observation it belongs to
}
\description{
This function samples the rejected proposals for the Truncated Mixture of Gaussian (TMoG) model
}
