% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ind_last_acc}
\alias{ind_last_acc}
\title{Index of last acceptance}
\usage{
ind_last_acc(num_a, left, vl)
}
\arguments{
\item{num_a}{total number of accepted samples}

\item{left}{number of observations without rejected samples}

\item{vl}{a vector of TRUE/FALSE indicating whether the samples are inside or outside the constraints}
}
\description{
This function provides the index of the observations with that is last accepted
}
