% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{indLastAcc}
\alias{indLastAcc}
\title{Index of last acceptance}
\usage{
indLastAcc(numA, left, vl)
}
\arguments{
\item{numA}{total number of accepted samples}

\item{left}{number of observations without rejected samples}

\item{vl}{a vector of TRUE/FALSE indicating whether the samples are inside or outside the constraints}
}
\description{
This function provides the index of the observations with that is last accepted
}
