% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_gibbs.R
\name{inf_constr_mixmod}
\alias{inf_constr_mixmod}
\title{Main function for inference}
\usage{
inf_constr_mixmod(mix, data_type, thr)
}
\arguments{
\item{mix}{Mixture type: one of "tmog" and "motg"}

\item{data_type}{one of "gauss", "beta", "bvgauss", "flowcyto", "crime"}

\item{thr}{vector of number of rejections per observations.}
}
\value{
samples from posterior distribution
}
\description{
This function evaluates the cumulative distribution of a univariate Gaussian distribution
}
