% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_algo.R
\name{inf_gvhd}
\alias{inf_gvhd}
\title{Gibbs sampling algorithm}
\usage{
inf_gvhd(K, ip_data, thr = Inf, burnIn, numIter, constr, mix, lower,
  upper, data_type)
}
\arguments{
\item{K}{maximum number of clusters}

\item{ip_data}{data matrix where columns are individual observations and rows are dimensions}

\item{thr}{maximum number of rejections per observations}

\item{burnIn}{number of burn-in for the chain. Must be less than number of iterations}

\item{numIter}{total number of iterations}

\item{constr}{function for the constraints}

\item{mix}{one of "tmogt" and "motg"}

\item{lower}{lower constraint for unit square}

\item{upper}{upper constriant for unit square}

\item{data_type}{one of "gauss", "beta", "flowcyto", "crime", "bvgauss"}
}
\value{
a list of results consisting of weights, mean, and covariance matrix for every iterations, along with the rejected proposals
}
\description{
This computes the Gibbs sampling procedure
}
