% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_gibbs.R
\name{inf_post_pred_mixmod}
\alias{inf_post_pred_mixmod}
\title{Main function for inference and posterior prediction}
\usage{
inf_post_pred_mixmod(mix, data_type, thr)
}
\arguments{
\item{mix}{Mixture type: one of "tmog" and "motg"}

\item{data_type}{one of "gauss", "beta", "bvgauss", "flowcyto", "crime"}

\item{thr}{vector of number of rejections per observations.}
}
\value{
result from Gibbs sampling, posterior prediction of test set, posterior likelihood, and time
}
\description{
This function implements Gibbs sampling and computes posterior predictive distribution of test set
}
