% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runGibbs.R
\name{infConstrMixMod}
\alias{infConstrMixMod}
\title{Main function for inference}
\usage{
infConstrMixMod(mix, Xdata, constr, K = 100, thr = 1, constrType,
  niwpr, ...)
}
\arguments{
\item{mix}{Mixture type: one of "tmog" and "motg"}

\item{Xdata}{p-by-N matrix where N is the total number of observations and p is the dimension}

\item{constr}{function to test whether observation belongs in a space. Must input Xdata and returns vector of TRUE/FALSE}

\item{K}{maximum number of clusters}

\item{thr}{vector of number of rejections per observations.}

\item{constrType}{one of "simple" and "complex". If "simple", follow by lower and upper bounds.}

\item{niwpr}{list of niw prior which includes m0, v0, a0, and b0}

\item{...}{pass lower and upper bounds of the constraint if type is "rectangle"}
}
\value{
samples from posterior distribution
}
\description{
This function evaluates the cumulative distribution of a univariate Gaussian distribution
}
