% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_gibbs.R
\name{Infernce_ConstrMixMod}
\alias{Infernce_ConstrMixMod}
\title{Main function for inference}
\usage{
Infernce_ConstrMixMod(mix, data_type, thr)
}
\arguments{
\item{mix}{one of "tmog" and "motg"}

\item{data_type}{one of "gauss", "beta", "bvgauss", "flowcyto", "crime"}

\item{thr}{vector of number of rejections per observations.}

\item{constr}{function to check constraints}
}
\value{
samples from posterior distribution
}
\description{
This function evaluates the cumulative distribution of a univariate Gaussian distribution
}
