% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_sampling.R
\name{mc_inter}
\alias{mc_inter}
\title{Monte Carlo integration function}
\usage{
mc_inter(constr, mn, cv, num_obs)
}
\arguments{
\item{constr}{a function that describes the domain or constraint of the observations}

\item{mn}{mean of the observations to be generated}

\item{cv}{covariance matrix of the observations to be generated}

\item{num_obs}{number of observations to be generated}
}
\value{
value of integrals
}
\description{
This function evaluates an integral using Monte Carlo integration.
}
\examples{
constr <- function(X, u1 = lower, u2 = upper) { X >= u1 & X <= u2 }
mc_inter(constr, 1/2, 0.1, 100)
                                
}
