% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niw.R
\name{nig}
\alias{nig}
\title{Normal-Inverse-Gamma function}
\usage{
nig(m0, v0, a0, b0)
}
\arguments{
\item{m0}{location parameter for the Normal distribution (real)}

\item{v0}{scale parameter for variance of the Normal distribution (real > 0)}

\item{a0}{shape parameter for the Inverse-Gamma distribution (real > 0)}

\item{b0}{rate parameter for the Inverse-Gamma distribution (real > 0)}
}
\value{
list with mean and variance
}
\description{
This function generates samples from a Normal-Inverse-Gamma density
}
\examples{
nig(0, 2, 2, 1)

}
