% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niw.R
\name{nig_post}
\alias{nig_post}
\title{Normal-Inverse-Gamma posterior density function}
\usage{
nig_post(m0, v0, a0, b0, SS)
}
\arguments{
\item{m0}{location parameter for the Normal distribution (real)}

\item{v0}{scale parameter for variance of the Normal distribution (real > 0)}

\item{a0}{shape parameter for the Inverse-Gamma distribution (real > 0)}

\item{b0}{rate parameter for the Inverse-Gamma distribution (real > 0)}

\item{SS}{a list which includes list(n = number of observations, mn = mean, S = covariance)}
}
\value{
a list with mean and variance
}
\description{
This function generates samples from the posterior density of a Normal-Inverse-Gamma funcion
}
\examples{
nig.post(0, 1, 2, 3, list(n = 2, mn = 1, S = 0.1))

}
