% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niw.R
\name{niw}
\alias{niw}
\title{Normal-Inverse-Wishart function}
\usage{
niw(mu0, lam, Phi, nu)
}
\arguments{
\item{lam}{scale parameter for the variance of the Gaussian distribution}

\item{Phi}{a positive definite scale matrix for the Inverse-Wishart distribution}

\item{nu}{degrees of freedom parameter for the Inverse-Whisart distribution}

\item{m0}{location parameter for the Gaussian distribution}
}
\value{
list with mean and covariance matrix
}
\description{
This function generates samples from a Normal-Inverse-Wishart density
}
\examples{
niw(rep(1/2,2), 1, diag(0.1, 2), 2)

}
