% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niw.R
\name{niwPost}
\alias{niwPost}
\title{Normal-Inverse-Wishart posterior density function}
\usage{
niwPost(mu0, lam, Phi, nu, SS)
}
\arguments{
\item{mu0}{location parameter for the Gaussian distribution}

\item{lam}{scale parameter for variance of the Gaussian distribution}

\item{Phi}{a positive definite scale matrix for the Inverse-Wishart distribution}

\item{nu}{degrees of freedom parameter for the Inverse-Whisart distribution}

\item{SS}{a list which includes list(n = number of observations, mn = mean, S = covariance matrix)}
}
\value{
a list with mean and covariance matrix
}
\description{
This function generates samples from the posterior density of a Normal-Inverse-Wishart funcion
}
\examples{
SS = list(n = 2, mn = rep(0.7, 2), S = diag(0.1,2))
niwPost(rep(1/2,2), 1, diag(0.1, 2), 2, SS)

}
