% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_pred_func.R
\name{norm_cdf}
\alias{norm_cdf}
\title{Cumulative distribution function}
\usage{
norm_cdf(mean, sd, lower, upper)
}
\arguments{
\item{mean}{mean of the distribution}

\item{sd}{standard deviation of the distribution}

\item{lower}{lower bound}

\item{upper}{upper bound}
}
\value{
the area between the two bounds
}
\description{
This function evaluates the cumulative distribution of a univariate Gaussian distribution
}
\examples{
constr <- function(X, u1 = lower, u2 = upper) { X >= u1 & X <= u2 }
mc_inter(constr, 1/2, 0.1, 100)

}
