% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_pred_func.R
\name{post_pred_motg}
\alias{post_pred_motg}
\title{Posterior predictive for univariate MoTG}
\usage{
post_pred_motg(res, test, K, pts, upper, lower, data_type, pr1, pr2)
}
\arguments{
\item{res}{list of results containing weight (wt), mean (mn), and covariance (cv) matrix for each cluster}

\item{test}{test set where columns are the individual observations and rows are the dimension}

\item{K}{number of clusters}

\item{pts}{grid points to draw the density distribution}

\item{upper}{upper bound of constraints, only use for unit square}

\item{lower}{lower bound of constraints, only use for unit square}

\item{data_type}{one of "gauss" and "beta"}

\item{pr1}{first parameter for the density}

\item{pr2}{second parameter for the density}
}
\value{
list of prediction at points, prediction of tests, along with error bars
}
\description{
This function evaluates the posterior predictive distribution for a univariate MoTG model
}
\examples{

}
