% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_pred_func.R
\name{post_pred_motg_mv}
\alias{post_pred_motg_mv}
\title{Posterior predictive for multivariate MoTG}
\usage{
post_pred_motg_mv(res, test, K, lower, upper, constr)
}
\arguments{
\item{res}{list of results containing weight (wt), mean (mn), and covariance (cv) matrix for each cluster}

\item{test}{test set where columns are the individual observations and rows are the dimension}

\item{K}{number of clusters}

\item{lower}{lower bound of constraints, only use for unit square}

\item{upper}{upper bound of constraints, only use for unit square}

\item{constr}{a function for the constraints}
}
\value{
list of prediction of tests
}
\description{
This function evaluates the posterior predictive distribution for a multivariate MoTG model
}
