% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_pred_func.R
\name{post_pred_motgt}
\alias{post_pred_motgt}
\title{Posterior predictive for MoTG}
\usage{
post_pred_motgt(res, test, K, pts, upper, lower, data_type, pr1, pr2)
}
\arguments{
\item{res}{list of results containing weight, mean, and covariance matrix for each cluster}

\item{test}{test set where columns are the individual observations and rows are the dimension}

\item{K}{number of clusters}

\item{pts}{grid points to draw the density distribution}
}
\value{
the area between the two bounds
}
\description{
This function evaluates the posterior predictive distribution for a univariate MoTG model
}
\examples{
constr <- function(X, u1 = lower, u2 = upper) { X >= u1 & X <= u2 }
mc_inter(constr, 1/2, 0.1, 100)

}
