% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_pred_func.R
\name{post_pred_tmog_mv}
\alias{post_pred_tmog_mv}
\title{Posterior predictive for multivariate TMoG}
\usage{
post_pred_tmog_mv(res, test, K, constr, thin_sc = 1)
}
\arguments{
\item{res}{list of results containing weight (wt), mean (mn), and covariance (cv) matrix for each cluster}

\item{test}{test set where columns are the individual observations and rows are the dimension}

\item{K}{number of clusters}

\item{constr}{a function for the constraints}

\item{thin_sc}{thinning for MCMC chain (default no thinning)}
}
\value{
list of prediction of tests
}
\description{
This function evaluates the posterior predictive distribution for a multivariate TMoG model
}
