% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postPredFuncs.R
\name{postPredDist}
\alias{postPredDist}
\title{Posterior predictive for univariate data}
\usage{
postPredDist(mix, constrType, res, test, pts, realDens, ...)
}
\arguments{
\item{mix}{one of "tmog" or "motg"}

\item{constrType}{one of "simple" and "complex". If "simple" include lower and upper bounds.}

\item{res}{list of results containing weight (wt), mean (mn), and covariance (cv) matrix for each cluster}

\item{test}{matrix of p-by-N where N are number of observations and p are number of dimension}

\item{pts}{grid points to draw the density distribution}

\item{realDens}{real density values at grid points}
}
\value{
list of prediction at points, prediction of tests, along with error bars
}
\description{
This function evaluates the posterior predictive distribution for univariate models
}
\examples{

}
