% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postPredFuncs.R
\name{postPredDistMv}
\alias{postPredDistMv}
\title{Posterior predictive for multivariate data}
\usage{
postPredDistMv(mix, constrType, res, test, constr, thinsc, ...)
}
\arguments{
\item{mix}{type of model. One of "tmog" or "motg"}

\item{constrType}{type of bounds. One of "simple" or "complex". If "simple", include lower and upper bounds for all dimension}

\item{res}{list of results containing weight (wt), mean (mn), and covariance (cv) matrix for each cluster}

\item{test}{matrix of p-by-N where p is the number of dimensions and N is the total number of observations}

\item{constr}{function which inputs test and outputs TRUE/FALSE}

\item{thinsc}{thinning of chain (default is 1, no thinning)}

\item{...}{input lower and upper bounds for "simple" constraint type}
}
\value{
list of prediction at points, prediction of tests, along with error bars
}
\description{
This function evaluates the posterior predictive distribution for univariate models
}
\examples{

}
