% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postPredFuncs.R
\name{postPredTmog}
\alias{postPredTmog}
\title{Posterior predictive for univariate TMoG}
\usage{
postPredTmog(res, test, pts, upper, lower, realDens)
}
\arguments{
\item{res}{list of results containing weight (wt), mean (mn), and covariance (cv) matrix for each cluster}

\item{test}{test set where columns are the individual observations and rows are the dimension}

\item{pts}{grid points to draw the density distribution}

\item{upper}{upper bound of constraints, only use for unit square}

\item{lower}{lower bound of constraints, only use for unit square}

\item{realDens}{real density values at points}
}
\value{
list of prediction at points, prediction of tests, along with error bars
}
\description{
This function evaluates the posterior predictive distribution for a univariate TMoG model
}
