% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postPredFuncs.R
\name{postPredTmogMv}
\alias{postPredTmogMv}
\title{Posterior predictive for multivariate TMoG}
\usage{
postPredTmogMv(res, test, constr, thinsc = 1, constrType, ...)
}
\arguments{
\item{res}{list of results containing weight (wt), mean (mn), and covariance (cv) matrix for each cluster}

\item{test}{test set where columns are the individual observations and rows are the dimension}

\item{constr}{a function for the constraints}

\item{thinsc}{thinning for MCMC chain (default no thinning)}

\item{constrType}{one of "simple" or "complex". Follow with lower and upper bounds if "simple".}
}
\value{
list of prediction of tests
}
\description{
This function evaluates the posterior predictive distribution for a multivariate TMoG model
}
