% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_func.R
\name{prob_norm}
\alias{prob_norm}
\title{Normalize probabilities}
\usage{
prob_norm(X)
}
\arguments{
\item{X}{vector of probabilities}
}
\value{
normalize vector of probabilities
}
\description{
This function normalizes a probability distribution
}
