% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{probNorm}
\alias{probNorm}
\title{Normalize probabilities}
\usage{
probNorm(X)
}
\arguments{
\item{X}{vector of probabilities}
}
\value{
normalize vector of probabilities
}
\description{
This function normalizes a probability distribution
}
