% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rejSummary}
\alias{rejSummary}
\title{Rejection summary}
\usage{
rejSummary(rejs, brks)
}
\arguments{
\item{rejs}{vector of rejected samples}

\item{brks}{break points of the bins to summarize the rejections}
}
\value{
number of samples in the bins
}
\description{
This function summarizes the value for the rejected samples
}
