% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateFuncs.R
\name{updateClustAssignMotg}
\alias{updateClustAssignMotg}
\title{Update assignment for univariate MoTG model}
\usage{
updateClustAssignMotg(K, ipdata, z, rej, N, params, thr)
}
\arguments{
\item{K}{max number of clusters}

\item{ipdata}{data matrix where columns are individual observations and rows are number of dimensions}

\item{z}{vector of cluster assignment}

\item{rej}{list of rejected proposals}

\item{N}{total number of observations}

\item{params}{list of parameters which include weights (wt), mean (mu), and covariance matrix (cv) for all clusters}

\item{thr}{maximum number of rejected samples per observations}
}
\value{
vector of new cluster assignments
}
\description{
This function updates the cluster assignment of each observation in the MoTG model
}
