% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateFuncs.R
\name{updateParamsComponent}
\alias{updateParamsComponent}
\title{Update components parameter}
\usage{
updateParamsComponent(X, mu, cv, niwpr, constr, thr = Inf)
}
\arguments{
\item{X}{data matrix where each column is independent observations and each row is the different dimension}

\item{mu}{mean of each cluster}

\item{cv}{covariance matrix of each cluster}

\item{niwpr}{Normal-inverse-wishart prior}

\item{constr}{function for the constraint}

\item{thr}{maximum number of rejections per observation}
}
\value{
list of the new mean, covariance matrix, number of observations in each cluster, number of rejections, and the rejected samples
}
\description{
This function update the parameters of each cluster for the MoTg model
}
