% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateFuncs.R
\name{updateParamsTmog}
\alias{updateParamsTmog}
\title{Update parameters for TMoG model}
\usage{
updateParamsTmog(K, alp, xRejs, wt, niwpr, zRejs)
}
\arguments{
\item{K}{max number of clusters}

\item{alp}{parameter for dirichlet process prior}

\item{xRejs}{data matrix of rejected proposals}

\item{wt}{weights of each cluster}

\item{niwpr}{normal-inverse-wishart prior}

\item{zRejs}{cluster assignments for rejected proposals}
}
\value{
list of new weights, mean, covariance matrix, number of rejected proposals, and number of observations in each cluster
}
\description{
This function updates the cluster parameters in the univariate TMoG model
}
