% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateFuncs.R
\name{updateWeights}
\alias{updateWeights}
\title{Update components weight}
\usage{
updateWeights(z, K, alp)
}
\arguments{
\item{z}{cluster assignment for every observations}

\item{K}{total number of clusters}

\item{alp}{hyperparameter of the dirichlet process prior}
}
\value{
vector of weights for every components
}
\description{
This function update the weights associated with each cluster
}
