% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_func.R
\name{updt_assgn_motg}
\alias{updt_assgn_motg}
\title{Update assignment for univariate MoTG model}
\usage{
updt_assgn_motg(K, ip_data, z, dm, rej, N, params, thr)
}
\arguments{
\item{K}{max number of clusters}

\item{ip_data}{data matrix where columns are individual observations and rows are number of dimensions}

\item{z}{vector of cluster assignment}

\item{dm}{number of dimensions}

\item{rej}{list of rejected proposals}

\item{N}{total number of observations}

\item{params}{list of parameters which include weights (wt), mean (mu), and covariance matrix (cv) for all clusters}

\item{thr}{maximum number of rejected samples per observations}
}
\value{
vector of new cluster assignments
}
\description{
This function updates the cluster assignment of each observation in the MoTG model
}
