% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_func.R
\name{updt_assgn_motgt}
\alias{updt_assgn_motgt}
\title{Update assignment for univariate MoTG model}
\usage{
updt_assgn_motgt(K, ip_data, N, dm, params, lower, upper)
}
\arguments{
\item{K}{max number of clusters}

\item{ip_data}{data matrix where columns are individual observations and rows are number of dimensions}

\item{N}{number of observations}

\item{dm}{number of dimensions}

\item{params}{list of parameters which include weights (wt), mean (mu), and covariance matrix (cv) for all clusters}

\item{lower}{lower bound of constraint (for unit square only)}

\item{upper}{upper bound of constraint (for unit square only)}
}
\value{
vector of new cluster assignments
}
\description{
This function updates the cluster assignment of each observation in the MoTG model
}
