% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_func.R
\name{updt_assgn_tmog}
\alias{updt_assgn_tmog}
\title{Update assignment for univariate TMoG model}
\usage{
updt_assgn_tmog(K, ip_data, N, dm, params)
}
\arguments{
\item{K}{max number of clusters}

\item{ip_data}{data matrix where columns are individual observations and rows are number of dimensions}

\item{N}{total number of observations}

\item{dm}{number of dimensions}

\item{params}{list of parameters which include weights (wt), mean (mu), and covariance matrix (cv) for all clusters}
}
\value{
vector of new cluster assignments
}
\description{
This function updates the cluster assignment of each observation in the univariate TMoG model
}
