% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_func.R
\name{updt_params_component}
\alias{updt_params_component}
\title{Update components parameter}
\usage{
updt_params_component(X, mu, cv, niw_p, constr, thr = Inf, dm)
}
\arguments{
\item{X}{data matrix where each column is independent observations and each row is the different dimension}

\item{mu}{mean of each cluster}

\item{cv}{covariance matrix of each cluster}

\item{niw_p}{Normal-inverse-wishart prior}

\item{constr}{function for the constraint}

\item{thr}{maximum number of rejections per observation}

\item{dm}{dimension of data}
}
\value{
list of the new mean, covariance matrix, number of observations in each cluster, number of rejections, and the rejected samples
}
\description{
This function update the parameters of each cluster for the MoTg model
}
