% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_func.R
\name{updt_params_tmog}
\alias{updt_params_tmog}
\title{Update parameters for TMoG model}
\usage{
updt_params_tmog(K, alp, x_rejs, dm, wt, niw_p, z_rejs)
}
\arguments{
\item{K}{max number of clusters}

\item{alp}{parameter for dirichlet process prior}

\item{x_rejs}{data matrix of rejected proposals}

\item{dm}{number of dimensions}

\item{wt}{weights of each cluster}

\item{niw_p}{normal-inverse-wishart prior}

\item{z_rejs}{cluster assignments for rejected proposals}
}
\value{
list of new weights, mean, covariance matrix, number of rejected proposals, and number of observations in each cluster
}
\description{
This function updates the cluster parameters in the univariate TMoG model
}
