% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_func.R
\name{updt_wghts}
\alias{updt_wghts}
\title{Update components weight}
\usage{
updt_wghts(z, K, alp)
}
\arguments{
\item{z}{cluster assignment for every observations}

\item{K}{total number of clusters}

\item{alp}{hyperparameter of the dirichlet process prior}
}
\value{
vector of weights for every components
}
\description{
This function update the weights associated with each cluster
}
