% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_pred_func.R
\name{edit_pred_prob}
\alias{edit_pred_prob}
\title{Edit underflow of predictive probability}
\usage{
edit_pred_prob(X)
}
\arguments{
\item{X}{vector of probabilities}
}
\description{
This function edits the underflow of the predictive probability function for truncated Gaussian
}
