% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoG.R
\name{gen_mog}
\alias{gen_mog}
\title{Mixture of Gaussian function}
\usage{
gen_mog(K = 2, params, N = 100, dm)
}
\arguments{
\item{K}{number of clusters (default 2)}

\item{params}{list with names: wt = weights, mu = mean, and cv = covariance}

\item{N}{number of random samples (default 100)}

\item{dm}{dimension}
}
\value{
list of the vector of random samples and the associated cluster assignment
}
\description{
This function generates a mixture of Gaussians samples.
}
\examples{
params = list(wt = c(0.5, 0.5), mu=c(0.1, 0.5), cv = c(0.1, 0.7))
gen_mog(K = 2, params, N = 100, dm = 2)
}
\keyword{Gaussians}
\keyword{Mixture}
\keyword{of}
