% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_gibbs.R
\name{inf_constr_mixmod}
\alias{inf_constr_mixmod}
\title{Main function for inference}
\usage{
inf_constr_mixmod(mix, X_data, constr, thr, constr_type, ...)
}
\arguments{
\item{mix}{Mixture type: one of "tmog" and "motg"}

\item{X_data}{p-by-N matrix where N is the total number of observations and p is the dimension}

\item{thr}{vector of number of rejections per observations.}

\item{constr_type}{one of "rectangle" and "complex". If rectangle, follow by lower and upper bounds.}

\item{...}{pass lower and upper bounds of the constraint if type is "rectangle"}
}
\value{
samples from posterior distribution
}
\description{
This function evaluates the cumulative distribution of a univariate Gaussian distribution
}
